refactor(interpreter): Split out `attr` and `style`
